// shesha-reactjs/src/apis/checkList.tsx
export interface CheckListItemModel {
  id?: string;
  /**
   * Item type (group = 1, two state = 2, tri state = 3), see Shesha.Domain.Enums.RefListCheckListItemType
   */
  itemType?: number;
  /**
   * Item name
   */
  name?: string | null;
  /**
   * Item description
   */
  description?: string | null;
  /**
   * If true, the user is able to add comments to this item/group
   */
  allowAddComments?: boolean;
  /**
   * Heading of the comments box
   */
  commentsHeading?: string | null;
  /**
   * Custom visibility of comments (javascript expression)
   */
  commentsVisibilityExpression?: string | null;
  /**
   * Child items
   */
  childItems?: CheckListItemModel[] | null;
}